requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# NLP
nltk>=3.8.0

# Machine Learning
scikit-learn>=1.3.0

# Statistical Analysis
statsmodels>=0.14.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
wordcloud>=1.9.3

# Data Loading
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0

# Utilities
tqdm>=4.66.0
